Add deprecation to arguments and commands (#38)
Adding new information to the help formatting methods in argparse proved difficult and as far as I can tell, I would have to either "monkey-patch" or do some messy inheritance. In the end, I just appended the deprecation to the help strings, which achieves the same thing, but it is less modular.

* Add dictionaries to store the names and reasons of deprecated keyword arguments and commands

* Append tag and message to help message

* Check for deprecated keyword arguments and commands when a command is run and log warning for each found